+++ xtract/xtract_helper.h
@@ -44,7 +44,7 @@ extern "C" {
 
 /** \brief Apply a window function to an array of length N
  *
- * \param *data a pointer to an array of floats
+ * \param *data a pointer to an array of doubles
  * \param N the number of elements in the array pointed to by *data
  * \param *argv a pointer to a window function as returned by xtract_make_window()
  * \param *result a pointer to the first element an array containing the windowed data
@@ -52,11 +52,11 @@ extern "C" {
  * It is up to the caller to generate and free the array containing the window, and to allocate and free memory of size N to hold the data pointed to by *result
  *
  */
-int xtract_windowed(const float *data, const int N, const void *argv, float *result);
+int xtract_windowed(const double *data, const int N, const void *argv, double *result);
 
 /** \brief Divides the array pointed to by *data into two subframes, and applies a given feature to each subframe, returning them in a single array pointed to by result
  *
- * \param *data an array of floats
+ * \param *data an array of doubles
  * \param N the number of elements in the array pointed by *data
  * \param feature an integer representing the feature to be applied to each subframe in data. This will be a value as given in the enumeration xtract_features_ (libxtract.h)
  * \param *argv a pointer to the argument vector to be passed to the feature extraction function as determined by feature
@@ -66,7 +66,7 @@ int xtract_windowed(const float *data, const int N, const void *argv, float *res
  * It is important to ensure that any _init_*() functions that are called in preparation for functions that are called on subframes are given the subframe size as 'N', and not the frame size. i.e. if xtract_features_from_subframes() is called with N=64, and feature=XTRACT_SPECTRUM, then xtract_init_fft() should be called with N=32.
  *
  */
-int xtract_features_from_subframes(const float *data, const int N, const int feature, const void *argv, float *result);
+int xtract_features_from_subframes(const double *data, const int N, const int feature, const void *argv, double *result);
 
 /** \brief Test whether a number is denormal */
 int xtract_is_denormal(double const d);
